Character sprite, memory icon updates and memory deck images (#23)
Added a basic character sprite attempting to follow instructions as https://github.com/daviscook477/BaseMod/wiki/Custom-Character-Animations. This has not been tested and I'm not even sure if it's the right scale, but I hope that calling jorbsWizard.scml instead of theDefaultAnimation.scml will work to produce an animated jorbsWizard sprite. 

Updated chastity, charity, humility, kindness, patience, envy, lust and pride to have less transparent icons to make them easier to see. Added one more image for each memory at the 48x48 pixel scale. Also added two thought-bubble images in the same directory to help see the memory you are currently in - called thoughtbubble32.png and thoughtbubble48.png. Each are larger than the named size, but they should be sized to fit an icon of the named size if it is placed at the center of the image. 

Added the 14 'recollection' deck card images, cut to be skill card portraits.