Track Wrath upgrades as a custom field
The previous solution, of implementing the CustomSavable interface in
CustomJorbsModCard, would only let Wrath upgrades be tracked on our
new cards. The main fail case is any gray attack card.

The new approach adds a custom field that's tracked per card. As Wrath
applies to a card, the custom field gets updated with a count of the
number of total applications to that card so far.

A separate class handles saving all per-card custom fields, of which
we only have the one currently. It is easily extended for more. The
current approach chooses to place its risk that the way cards are
stored in the save file stay in order. This is in contrast to placing
the risk on patch locations that would guarantee ordering better but
need to be updated more often in response to base game changes.