Persist Wrath modifiers across play sessions (#106)
* Persist Wrath modifiers across play sessions

Cards save their ID, number of upgrades, and their misc field by
default. To save how Wrath modifies a card's base damage, we
implement the CustomSavable interface on all of our cards and
introduce a new class to track persistent changes. Currently the only
value is how much Wrath modifies a card. Persistent changes are
re-applied to cards upon loading.

* Track Wrath upgrades as a custom field

The previous solution, of implementing the CustomSavable interface in
CustomJorbsModCard, would only let Wrath upgrades be tracked on our
new cards. The main fail case is any gray attack card.

The new approach adds a custom field that's tracked per card. As Wrath
applies to a card, the custom field gets updated with a count of the
number of total applications to that card so far.

A separate class handles saving all per-card custom fields, of which
we only have the one currently. It is easily extended for more. The
current approach chooses to place its risk that the way cards are
stored in the save file stay in order. This is in contrast to placing
the risk on patch locations that would guarantee ordering better but
need to be updated more often in response to base game changes.

* Cleanup and polish

Rename fields to separate from base game's use of upgrade.
Restore battle-instance cards to receive the Wrath field update.
Patch AbstractCard to copy our data in the stat-equivalent method.

* fix off-by-one in makeStatEquivalentCopy patch

* update name of wrathEffectCount field to make it more clear during debugging