Fix card interaction with Nilry's Codex
A number of random card generation methods eliminate cards with
healing effects from the eligible set. We honor the letter and the
spirit of that filtering by tagging our direct healing cards and,
additionally, eliminating cards with a new tag indicating any other
persistent positive effects from being generated as well.

Future cards must be careful to use these tags in order to
indicate when they are not eligible to be generated as a random card.

Resolves #70.