Implemented Cull Rare card: Consumption X: Deal X damage, gain max hp equal to damage dealt. Exhaust.
Damage and MaxHP gain scale with all damage modifiers (strength, vuln, pen nib etc.). MaxHP gain is equal to the actual hp loss of the target, and as such is blocked by shield, and capped by the actual HP of the target.

One note, I couldn't find a clean way to add dynamic text, since the damage is calculated in use() based on energy spent. Feel free to give implementation suggestions as feedback.