Initial implementation of Exert mechanic
In the current implementation, if you want a card to be exerted when used, you have to set the EntombedField of the masterDeckEquivalent to true, as well as set the new CustomJorbsModCard.exert property to true. Cards with this property are checked at the start of the act, setting their Entombed property to false.
Currently there is no way to distinguish an exerted card from an exhausted card in the exhaust pile. Also, a card has no 'exerted' property to hook onto, it is all managed using the entombed mechanics. That might be a welcome addition.

This PR includes the card 'Frustration', made for testing the exert mechanics. It is fully functional and can be reviewed on its own merits.