Implements Cull card Equinox [cost = current act]: Deal damage equal to current floor number. Exhaust.
I couldn't think of an elegant way to establish the dynamic cost of this card, so I had to implement 2 new basemod subscribers, one that updates the cost whenever you change acts, and one that determines the correct cost when loading a game. If there's an easier way to approach this, please let me know!