Implements Cull card Gathering Evil (#467)
* Implements Cull card Gathering Evil
1: Gain 1 Strength and 1 Vulnerable for every curse played this fight. Exhaust.
Upgrade: Entomb: When deck shuffles

Added an OnShuffle subscriber. Currently, onShuffle is only called for cards in the Exhaust pile, since this is the only card that currently requires this subscriber, and only when it is exhausted. I'm not sure if this is wise, or if it is better to call it for each card anywhere.
Also, I interpreted this upgrade to mean: Exhume when your deck shuffles. It doesn't also entomb the card, since that seems like even more of a change from the unupgraded card.

* Added check on magicNumber > 0, to prevent applying 0 stacks of two powers.

Co-authored-by: Dan Bjorge <dan@dbjorge.net>