Implements Cull card Pollute (#471)
* Implements Cull card Pollute.
1: All cards in your hand become Curses for the rest of combat. Exhaust

Went for the simplest execution, it simply changes each card's type to Curse. This makes them no longer count as their previous type, so they won't trigger Gremlin Nob, won't count for Kunai etc. Also, it changes their frame to the Skill frame, which changes the cropping of the art. Not sure how to fix this.

* Fixed upgrade method

Co-authored-by: Dan Bjorge <dan@dbjorge.net>