Cull mechanic exert (#474)
* Attempt at making Exert mechanic, with initial card Frustration.
Not working currently, exerted status not stored properly.

* Initial implementation of Exert mechanic

In the current implementation, if you want a card to be exerted when used, you have to set the EntombedField of the masterDeckEquivalent to true, as well as set the new CustomJorbsModCard.exert property to true. Cards with this property are checked at the start of the act, setting their Entombed property to false.
Currently there is no way to distinguish an exerted card from an exhausted card in the exhaust pile. Also, a card has no 'exerted' property to hook onto, it is all managed using the entombed mechanics. That might be a welcome addition.

This PR includes the card 'Frustration', made for testing the exert mechanics. It is fully functional and can be reviewed on its own merits.

* Corrected after review:
- adds two SpireFields: SelfExert and Exerted
- cards with SelfExertField = true will automatically be exhausted when used, and ExertedField gets set to true
- ExertedPatch sets cards with Exerted = true to start in Exhaust pile.
- Exerted no longer uses Entombed field, so that Exerting cards with Entombed does not override their Entombed state
- Frustration reworded to match new Retain: / Exert: template
- Exert keyword now uses correct coloring and capitalization
- MakeStatEquivalentCopyPatch now also copies Exerted state for cards that are not CustomJorbsModCards

* Corrected after review:
- attempted corrections in hook pattern
- removes explicit Exhaust from Frustration
- adds EntombedField and ExertedField to CardSaveData

* Corrected after review:
- further corrections in hook usage
- corrects index references in CardSaveData
- adds reset of ExertedField in JorbsMod::receiveStartAct

* Corrected after review:
- removed implement of StartActSubscriber on CustomJorbsModCard

* Changelog for Frustratoin

Co-authored-by: Dan Bjorge <dan@dbjorge.net>