Implements Cull card Rebuttal 1: Retain: gain 2 thorns while in hand. Exert: gain 4 thorns.
Perhaps there's a smarter implementation possible. This one applies the thorns when it is first retained. Then, all causes for removal from hand (use, discard, exhaust) trigger a loss of these thorns. There is a risk that there's other causes of getting the card from your hand that do not trigger the thorns removal, though, but I'm not sure how to handle the 'while in your hand' any other way.